fixed changeDetection bug

+++ backend/pkg/api/job/job.go
@@ -1,49 +1,51 @@
 package job
+
 import (
 	"PPA"
-	"fmt"
-	"context"
+	//"fmt"
 	"bytes"
+	"context"
 	"net/http"
 	"time"
+
 	"github.com/gin-gonic/gin"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
 const (
-	NotFound = http.StatusNotFound
-	Conflict = http.StatusConflict
-	Collection = "jobs"
+	NotFound         = http.StatusNotFound
+	Conflict         = http.StatusConflict
+	Collection       = "jobs"
 	ClientCollection = "clients"
-	EventTag = "m"
+	EventTag         = "m"
 )
 
 var OidNotFound = PPA.NewAppError(NotFound, "Does not exist")
 var JobNotFound = PPA.NewAppError(NotFound, "Does not exist")
 
 type Update struct {
-	Title string
-	ClientName string
-	ClientPhone string
+	Title          string
+	ClientName     string
+	ClientPhone    string
 	AssignedWorker primitive.ObjectID
-	CarInfo string
-	Notes string
-	StartTime string
-	EndTime string
-	Tag string
+	CarInfo        string
+	Notes          string
+	StartTime      string
+	EndTime        string
+	Tag            string
 	//Color *PPA.CalendarMeta
-	PrimaryColor string
+	PrimaryColor   string
 	SecondaryColor string
 }
 
 type ClientUpdate struct {
-	Name string
+	Name  string
 	Phone string
 }
 
 func (j Job) Create(c *gin.Context, req PPA.Job, editor PPA.Editor) (*PPA.Job, error) {
-	duration := time.Now().Add(5*time.Second)
+	duration := time.Now().Add(5 * time.Second)
 	ctx, cancel := context.WithDeadline(c.Request.Context(), duration)
 	defer cancel()
 
@@ -56,7 +58,8 @@ func (j Job) Create(c *gin.Context, req PPA.Job, editor PPA.Editor) (*PPA.Job, e
 	const matched int = 0
 	if bytes.Compare([]byte(req.AssignedWorker.Hex()), []byte(primitive.NilObjectID.Hex())) != matched {
 		// update tinter
-		tinter, tinterErr := j.tdb.ViewById(j.db, ctx, req.AssignedWorker); if tinterErr == nil {
+		tinter, tinterErr := j.tdb.ViewById(j.db, ctx, req.AssignedWorker)
+		if tinterErr == nil {
 			workerAssignChanged = true
 			cpyTinter = tinter
 			_ = j.tdb.AssignJobId(j.db, ctx, tinter.JobsCollection, req.ID)
@@ -64,7 +67,8 @@ func (j Job) Create(c *gin.Context, req PPA.Job, editor PPA.Editor) (*PPA.Job, e
 			return nil, PPA.NewAppError(NotFound, "Tinter does not exist")
 		}
 	}
-	created, err := j.jdb.Create(j.db, ctx, &req); if err != nil {
+	created, err := j.jdb.Create(j.db, ctx, &req)
+	if err != nil {
 		return nil, err
 	}
 
@@ -88,11 +92,11 @@ func (j Job) Create(c *gin.Context, req PPA.Job, editor PPA.Editor) (*PPA.Job, e
 
 	if !j.clientExists(ctx, created.ClientPhone) {
 		// create the client
-		if _, err := j.createClient(ctx, &PPA.Client {
-			ID: primitive.NilObjectID,
-			Name: created.ClientName,
+		if _, err := j.createClient(ctx, &PPA.Client{
+			ID:    primitive.NilObjectID,
+			Name:  created.ClientName,
 			Phone: created.ClientPhone,
-			Jobs: []primitive.ObjectID{},
+			Jobs:  []primitive.ObjectID{},
 		}, editor); err != nil {
 			return nil, err
 		}
@@ -106,19 +110,20 @@ func (j Job) Create(c *gin.Context, req PPA.Job, editor PPA.Editor) (*PPA.Job, e
 	return created, nil
 }
 
-func (j Job) ViewById( c *gin.Context, id string) (*PPA.Job, error) {
-	duration := time.Now().Add(5*time.Second)
+func (j Job) ViewById(c *gin.Context, id string) (*PPA.Job, error) {
+	duration := time.Now().Add(5 * time.Second)
 	ctx, cancel := context.WithDeadline(c.Request.Context(), duration)
 	defer cancel()
 
-	oid, err := primitive.ObjectIDFromHex(id); if err != nil {
+	oid, err := primitive.ObjectIDFromHex(id)
+	if err != nil {
 		return nil, OidNotFound
 	}
 	return j.jdb.ViewById(j.db, ctx, oid)
 }
 
 func (j Job) List(c *gin.Context, opts PPA.BulkFetchOptions) (*[]PPA.Job, error) {
-	duration := time.Now().Add(5*time.Second)
+	duration := time.Now().Add(5 * time.Second)
 	ctx, cancel := context.WithDeadline(c.Request.Context(), duration)
 	defer cancel()
 
@@ -126,15 +131,17 @@ func (j Job) List(c *gin.Context, opts PPA.BulkFetchOptions) (*[]PPA.Job, error)
 }
 
 func (j Job) Delete(c *gin.Context, id string, editor PPA.Editor) error {
-	duration := time.Now().Add(5*time.Second)
+	duration := time.Now().Add(5 * time.Second)
 	ctx, cancel := context.WithDeadline(c.Request.Context(), duration)
 	defer cancel()
 
-	oid, err := primitive.ObjectIDFromHex(id); if err != nil {
+	oid, err := primitive.ObjectIDFromHex(id)
+	if err != nil {
 		return OidNotFound
 	}
 
-	job, err := j.jdb.ViewById(j.db, ctx, oid); if err != nil {
+	job, err := j.jdb.ViewById(j.db, ctx, oid)
+	if err != nil {
 		return JobNotFound
 	}
 
@@ -142,8 +149,9 @@ func (j Job) Delete(c *gin.Context, id string, editor PPA.Editor) error {
 		return err
 	}
 
-	tinter, tinterErr := j.tdb.ViewById(j.db, ctx, job.AssignedWorker); if tinterErr == nil {
-	_ = j.tdb.RemoveJobId(j.db, ctx, tinter.JobsCollection, job.AssignedWorker)
+	tinter, tinterErr := j.tdb.ViewById(j.db, ctx, job.AssignedWorker)
+	if tinterErr == nil {
+		_ = j.tdb.RemoveJobId(j.db, ctx, tinter.JobsCollection, job.AssignedWorker)
 		tinter.AppendLog(j.eventLogger.LogDeleted(ctx, editor))
 		j.tdb.LogEvent(j.db, ctx, tinter)
 	} else {
@@ -156,22 +164,24 @@ func (j Job) Delete(c *gin.Context, id string, editor PPA.Editor) error {
 }
 
 func (j Job) Update(c *gin.Context, req Update, id string, editor PPA.Editor) (*PPA.Job, error) {
-	duration := time.Now().Add(5*time.Second)
+	duration := time.Now().Add(5 * time.Second)
 	ctx, cancel := context.WithDeadline(c.Request.Context(), duration)
 	defer cancel()
 
-	oid, err := primitive.ObjectIDFromHex(id); if err != nil {
+	oid, err := primitive.ObjectIDFromHex(id)
+	if err != nil {
 		return nil, OidNotFound
 	}
 
-	oldJob, err := j.jdb.ViewById(j.db, ctx, oid); if err != nil {
+	oldJob, err := j.jdb.ViewById(j.db, ctx, oid)
+	if err != nil {
 		return nil, err
 	}
 
-	client, err := j.cdb.ViewByPhone(j.db, ctx, oldJob.ClientPhone); if err != nil {
+	client, err := j.cdb.ViewByPhone(j.db, ctx, oldJob.ClientPhone)
+	if err != nil {
 		return nil, err
 	}
-
 
 	var workerAssignChanged bool = false
 	var cpyTinter *PPA.Tinter
@@ -179,7 +189,8 @@ func (j Job) Update(c *gin.Context, req Update, id string, editor PPA.Editor) (*
 	if bytes.Compare([]byte(req.AssignedWorker.Hex()), []byte(primitive.NilObjectID.Hex())) != matched {
 		workerAssignChanged = true
 		// update tinter
-		tinter, tinterErr := j.tdb.ViewById(j.db, ctx, req.AssignedWorker); if tinterErr == nil {
+		tinter, tinterErr := j.tdb.ViewById(j.db, ctx, req.AssignedWorker)
+		if tinterErr == nil {
 			cpyTinter = tinter
 			// remove old OID
 			_ = j.tdb.RemoveJobId(j.db, ctx, tinter.JobsCollection, oldJob.AssignedWorker)
@@ -199,36 +210,36 @@ func (j Job) Update(c *gin.Context, req Update, id string, editor PPA.Editor) (*
 		}
 	}
 
-	if err := j.jdb.Update(j.db, ctx, oid, &PPA.Job {
-		ID: primitive.NilObjectID,
-		Title: req.Title,
-		ClientName: req.ClientName,
-		ClientPhone: req.ClientPhone,
+	if err := j.jdb.Update(j.db, ctx, oid, &PPA.Job{
+		ID:             primitive.NilObjectID,
+		Title:          req.Title,
+		ClientName:     req.ClientName,
+		ClientPhone:    req.ClientPhone,
 		AssignedWorker: req.AssignedWorker,
-		CarInfo: req.CarInfo,
-		Notes: req.Notes,
-		Tag: req.Tag,
+		CarInfo:        req.CarInfo,
+		Notes:          req.Notes,
+		Tag:            req.Tag,
 		//Color: req.Color,
-		PrimaryColor: req.PrimaryColor,
+		PrimaryColor:   req.PrimaryColor,
 		SecondaryColor: req.SecondaryColor,
-		StartTime: req.StartTime,
-		EndTime: req.EndTime,
-
+		StartTime:      req.StartTime,
+		EndTime:        req.EndTime,
 	}); err != nil {
 		return nil, err
 	}
 
 	if client.Name != req.ClientName || client.Phone != req.ClientPhone {
-		if err = j.cdb.Update(j.db, ctx, client.ID, &PPA.Client {
-			Name: req.ClientName,
+		if err = j.cdb.Update(j.db, ctx, client.ID, &PPA.Client{
+			Name:  req.ClientName,
 			Phone: req.ClientPhone,
 		}); err != nil {
 			return nil, err
 		}
-		j.updateJobs(ctx, client.Jobs, oid, ClientUpdate{ Name: req.ClientName, Phone: req.ClientPhone }, editor)
+		j.updateJobs(ctx, client.Jobs, oid, ClientUpdate{Name: req.ClientName, Phone: req.ClientPhone}, editor)
 	}
 
-	updated, err := j.jdb.ViewById(j.db, ctx, oid); if err != nil {
+	updated, err := j.jdb.ViewById(j.db, ctx, oid)
+	if err != nil {
 		return nil, PPA.InternalError
 	}
 
@@ -242,7 +253,8 @@ func (j Job) Update(c *gin.Context, req Update, id string, editor PPA.Editor) (*
 
 	// fetch the old job's tinter for logging
 	var loggableOldTinter *PPA.LoggableTinter
-	oldTinter, oldTErr := j.tdb.ViewById(j.db, ctx, oldJob.AssignedWorker); if oldTErr != nil {
+	oldTinter, oldTErr := j.tdb.ViewById(j.db, ctx, oldJob.AssignedWorker)
+	if oldTErr != nil {
 		loggableOldTinter = nil
 	} else {
 		loggableOldTinter = oldTinter.Loggable()
@@ -283,7 +295,8 @@ func (j Job) createClient(ctx context.Context, client *PPA.Client, editor PPA.Ed
 		client.ID = primitive.NewObjectID()
 	}
 
-	created, err := j.cdb.Create(j.db, ctx, client); if err != nil {
+	created, err := j.cdb.Create(j.db, ctx, client)
+	if err != nil {
 		return nil, err
 	}
 
@@ -302,7 +315,8 @@ func (j Job) clientOidExists(ctx context.Context, oid primitive.ObjectID) bool {
 }
 
 func (j Job) attatchJobToClient(ctx context.Context, phone string, job *PPA.Job, editor PPA.Editor) error {
-	client, err := j.cdb.ViewByPhone(j.db, ctx, phone); if err != nil {
+	client, err := j.cdb.ViewByPhone(j.db, ctx, phone)
+	if err != nil {
 		return err
 	}
 
@@ -315,18 +329,20 @@ func (j Job) attatchJobToClient(ctx context.Context, phone string, job *PPA.Job,
 }
 
 func (j Job) updateJobs(ctx context.Context, oids []primitive.ObjectID, currOID primitive.ObjectID, update ClientUpdate, editor PPA.Editor) {
-	if len(oids) <= 0 { return } // Note that len(nil) is 0
+	if len(oids) <= 0 {
+		return
+	} // Note that len(nil) is 0
 
 	for _, oid := range oids {
 		const matched int = 0
 		if bytes.Compare([]byte(oid.Hex()), []byte(currOID.Hex())) != matched {
 			oldDoc, _ := j.jdb.ViewById(j.db, ctx, oid)
 
-			if err := j.jdb.Update(j.db, ctx, oid, &PPA.Job {
-				ClientName: update.Name,
+			if err := j.jdb.Update(j.db, ctx, oid, &PPA.Job{
+				ClientName:  update.Name,
 				ClientPhone: update.Phone,
 			}); err != nil {
-			// do nothing or fail, depends
+				// do nothing or fail, depends
 			}
 
 			newDoc, _ := j.jdb.ViewById(j.db, ctx, oid)
